+++ example.py
@@ -3,9 +3,10 @@ import os
 import pprint
 import time
 
+from notion_database.children import Children
 from notion_database.database import Database
 from notion_database.page import Page
-from notion_database.properties import Properties, Children
+from notion_database.properties import Properties
 
 try:
     from dotenv import load_dotenv
@@ -64,8 +65,48 @@ for i in D.result["results"]:
     PROPERTY.set_email("email", "test@test.com")
     PROPERTY.set_phone_number("phone", "010-0000-0000")
 
+    # Children block
+    logger.debug("Set Children block")
     children = Children()
-    children.set_body("hello world!")
+    # children.set_body("hello world!")  # deprecated. set_body -> set_paragraph
+    children.set_paragraph("set_paragraph")
+
+    children.set_heading_1("set_heading_1")
+    children.set_heading_2("set_heading_2")
+    children.set_heading_3("set_heading_3")
+
+    children.set_callout("set_callout")
+
+    children.set_quote("set_quote")
+
+    children.set_bulleted_list_item("set_bulleted_list_item")
+
+    children.set_numbered_list_item("first set_numbered_list_item")
+    children.set_numbered_list_item("second set_numbered_list_item")
+
+    children.set_to_do("set_to_do", checked=True)
+    children.set_to_do("set_to_do", checked=False)
+
+    children.set_toggle("set_toggle", children_text="WOW!")
+
+    children.set_code("set_code")
+    children.set_code("const a = 1", lang="javascript")
+    children.set_code("print(\"hello world!\")", lang='python')
+
+    children.set_embed("https://www.google.com")
+
+    children.set_external_image("https://github.githubassets.com/images/modules/logos_page/Octocat.png")
+    children.set_external_video("http://download.blender.org/peach/trailer/trailer_480p.mov")
+    children.set_external_file("https://github.com/microsoft/ML-For-Beginners/raw/main/pdf/readme.pdf")
+    children.set_external_pdf("https://github.com/microsoft/ML-For-Beginners/blob/main/pdf/readme.pdf")
+
+    children.set_bookmark("https://www.google.com")
+
+    children.set_equation("e=mc^2")
+
+    children.set_divider()
+    children.set_table_of_contents()
+    children.set_breadcrumb()
 
     # Create Page
     logger.debug("Create Page")